+++ tests/cases/case_8/case_8.h
@@ -1,1 +1,1 @@
-void sub_8(double x);
+void sub_8(short x);
