+++ tests/cases/case_9/case_9.h
@@ -1,1 +1,1 @@
-void sub_9(double x);
+void sub_9(long x);
